requirements_streamlit.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
mysql-connector-python>=8.3.0
python-dotenv>=1.0.1
requests>=2.31.0
